servers/src/pyproject.toml
[project]
name = "book-code-servers"
version = "0.1.0"
description = "MCP servers for AI agents"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.169.0",
    "mcp>=1.9.0",
    "python-dotenv>=1.0.0",
]
